スタイリングなどはmain.pyのコードが長くならないように、インポートするようにしました。
